chore: split test_indicators.py into category-grouped files (#545)
The 562-line test_indicators.py was the suite's largest file and grew with
every new indicator. Split it, by the frontend registry's category taxonomy,
into three cohesive files:

- test_indicators_technical.py  — rsi, sma, macd, nefi, cmf (15)
- test_indicators_volatility.py — atr, atr_pct, adx, chop, vnr (32)
- test_indicators_core.py       — registry/compute cross-cutting tests (3)

Pure move: each test's body is byte-identical (extracted via ast, not
retyped). The local _series_from_returns helper moved to tests/helpers.py as
the public make_series_from_returns (aliased on import, so test bodies are
unchanged). Same 50 tests collected; full suite still 675 passed; ruff clean.
